🎨 style(Card.tsx): add support for variant prop to Card component The Card component now supports a variant prop that can be used to change the border and shadow styles of the card. The default variant is 'default', which has a border and shadow, while the 'noBorder' variant has no border or shadow. This change improves the flexibility of the Card component and allows for more customization options.
🚀 feat(EventHero.tsx): remove buyFunction and buyText props, remove Button component
The buyFunction and buyText props have been removed from the EventHero component, as they are not needed. The Button component has also been removed, as it is not used anymore. This change simplifies the EventHero component and removes unnecessary code.

🚀 feat(Event.tsx): remove Button component, add QrCode icon
The Button component has been removed from the Event component, as it is not needed anymore. The QrCode icon has