🐛 fix(page.tsx): redirect to root path instead of '/cart' when user is not KYC validated to improve user experience
The redirect was changed from '/cart' to '/' in both `cancelled/page.tsx` and `success/page.tsx` files. This change ensures that when a user is not KYC validated, they are redirected to the root path instead of the '/cart' page. This improves the user experience by preventing them from accessing the cart pages when their KYC validation is not complete.