🔧 chore(migrations): update nftTransfer table schema in down.sql and up.sql files
📦 feat(gql): add NftTransfer fragment and mutation/query files for nftTransfer

In the migrations, the nftTransfer table schema was updated in both the down.sql and up.sql files. The changes include adding columns for eventPassId, packId, and packAmount, as well as adding constraints and comments for the table and columns.

In the gql folder, new files were added for the NftTransfer fragment and mutation/query. The NftTransfer fragment includes the fields for the nftTransfer table, and the mutation/query files provide functionality for upserting and retrieving nftTransfer data.