🚀 feat(cart-api): add getAllPassesCart function to retrieve all passes in the cart
📝 This commit adds a new function called `getAllPassesCart` to the `cart-api` feature. This function retrieves all passes in the cart and organizes them by organizer and event. It takes an optional parameter `userPassPendingOrders` which is an array of user pass pending orders. If this parameter is provided, the function will use it to generate the cart. Otherwise, it will fetch the cart from the pass cache.

The function iterates over the `userPassPendingOrders` array and groups the orders by organizer and event. It returns an object `allPassesCart` which has the following structure:

```
{
  [organizerSlug]: {
    [eventSlug]: [orders]
  }
}
```

If the `userPassPendingOrders` parameter is not provided, the function fetches the cart from the pass cache using the `passCache.getAllPassesCart` method.

There is a TODO comment in the code suggesting the addition of a cache validation process.